Crate restructure

The restructure crate provides functionalities to match regexp patterns into struct fields. This crate only builds on rust-nightly as of now.

To use this, add the following to your Cargo.toml.

 [dependencies]
 restructure = "0.1.1"

and add this to your crate root:

 #![feature(cell_extras)]
 extern crate regex;

Macros

regexify!
Create a struct with regex patterns and implements RegexStruct trait on it.

Structs

Restruct
Contains the fill method used to fill the struct with fields

Traits

RegexStruct
This trait allows you to match struct fields with regexp